bluetooth: Fix compilation warning issue after enabling AVRCP in Zephyr.
bug: v/68551
